==> scripts_notebooks/07-data_summary.ipynb <==
import numpy as np
import pandas as pd
import joblib
from pathlib import Path

DATA_PATH = Path("data")

# Iterate over all the folders in the data folder
df_l = []
for folder in DATA_PATH.iterdir():
    if not folder.is_dir():
        continue
    config_files = list(folder.glob("*config*"))
    row_dict = {}
    #Loading any one file cause its all the same
    config = np.load(config_files[0], allow_pickle=True).item()
    row_dict['dataset_name'] = config['dataset_name']
    row_dict['type'] = "classification" if config["regression"]==0 else "regression"
    row_dict['has_categorical_features'] = "data__categorical" in config and config["data__categorical"]==1
    row_dict['n_folds'] = len(config_files)
    #Loading any one train file
    train = np.load(folder/"x_train_fold_0.npy", allow_pickle=True)
    row_dict['n_features'] = train.shape[1]
    row_dict['n_train'] = train.shape[0]
    val = np.load(folder/"x_val_fold_0.npy", allow_pickle=True)
    row_dict['n_val'] = val.shape[0]
    test = np.load(folder/"x_test_fold_0.npy", allow_pickle=True)
    row_dict['n_test'] = test.shape[0]
    df_l.append(row_dict)

data_summary_df = pd.DataFrame(df_l)
data_summary_df

data_summary_df[data_summary_df.n_features>=20]

data_summary_df.to_parquet("data/data_summary.parquet")

# data_summary_df = pd.read_parquet("data/data_summary.parquet")
